Sprint 15, review 2

+++ src/main/java/ru/practicum/shareit/booking/service/BookingService.java
@@ -1,5 +1,6 @@
 package ru.practicum.shareit.booking.service;
 
+import org.springframework.data.domain.Page;
 import ru.practicum.shareit.booking.dto.BookingDtoRequest;
 import ru.practicum.shareit.booking.dto.BookingDtoResponse;
 
@@ -13,7 +14,7 @@ public interface BookingService {
 
     BookingDtoResponse findById(long userId, long id);
 
-    List<BookingDtoResponse> findAll(long userId, String state, Long from, Integer size);
+    List<BookingDtoResponse> findAll(long userId, String state, Integer from, Integer size);
 
     List<BookingDtoResponse> findAllByOwner(long userId, String state, Integer from, Integer size);
